requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/base.txt requirements/base.in
#
alembic==1.4.2            # via flask-migrate
certifi==2020.4.5.2       # via python-telegram-bot
cffi==1.14.0              # via cryptography
click==7.0                # via flask
cryptography==2.9.2       # via python-telegram-bot
decorator==4.4.2          # via python-telegram-bot
flask-login==0.5.0        # via -r requirements/base.in
flask-migrate==2.5.3      # via -r requirements/base.in
flask-sqlalchemy==2.4.1   # via -r requirements/base.in, flask-migrate
flask==1.1.2              # via -r requirements/base.in, flask-login, flask-migrate, flask-sqlalchemy
future==0.18.2            # via python-telegram-bot
gunicorn==20.0.4          # via -r requirements/base.in
itsdangerous==1.1.0       # via flask
jinja2==2.11.1            # via flask
mako==1.1.1               # via alembic
markupsafe==1.1.1         # via jinja2, mako
psycopg2==2.8.5           # via -r requirements/base.in
pycparser==2.20           # via cffi
python-dateutil==2.8.1    # via alembic
python-editor==1.0.4      # via alembic
python-telegram-bot==12.7  # via -r requirements/base.in
six==1.14.0               # via cryptography, python-dateutil
sqlalchemy==1.3.13        # via alembic, flask-sqlalchemy
tornado==6.0.4            # via python-telegram-bot
werkzeug==1.0.0           # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
